script: Add replay.py python script
This adds a new python script, replay.py, in scripts directory.
The usage of the script is as follows:

The below example shows replay output in record time and it almost looks
like 'uftrace replay' output.  It also prints return values if it has.

  $ uftrace record -s scripts/replay.py -F main -R .*@retval tests/t-abc
  # DURATION    TID     FUNCTION
              [25794] | main() {
              [25794] |   a() {
              [25794] |     b() {
              [25794] |       c() {
              [25794] |         getpid() {
    11.037 us [25794] |         } = 25794; /* getpid */
    44.752 us [25794] |       } = 25794; /* c */
    70.924 us [25794] |     } = 25795; /* b */
    98.191 us [25794] |   } = 25794; /* a */
   124.329 us [25794] | } = 0; /* main */

The next execution output shows how to run 'uftrace script' command for
recorded data.  As of now 'uftrace script' cannot support args/retval,
but the output also looks similar to 'uftrace replay'.

  $ uftrace script -s scripts/replay.py
  # DURATION    TID     FUNCTION
              [25794] | main() {
              [25794] |   a() {
              [25794] |     b() {
              [25794] |       c() {
              [25794] |         getpid() {
    11.037 us [25794] |         }; /* getpid */
    44.752 us [25794] |       }; /* c */
    70.924 us [25794] |     }; /* b */
    98.191 us [25794] |   }; /* a */
   124.329 us [25794] | }; /* main */

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>