build: Handle if system does not have libpython2.7.so
If system doesn't have libpython2.7.so installed, it has to exclude some
files and routines that are related to python features.
This patch handles such cases and do not support script related option
and command if system doesn't have libpython2.7.so.

Script option and command are independent from python features, but
those are also excluded until other script support is added.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>