plthook: Fix size format in debug message.
Since the type of sym->addr is uint64_t, its print format has to be
"%llx" instead of "%lx".  Because of this problem, the following
argument sym->name is not shown properly in 32-bit arch.

Before:
  # uftrace record --debug-domain plthook:3 t-abc.arm
      ...
  plthook: [idx:    2] enter 10480: (null) (mod: 14b9e38)
  plthook: [idx:    2] exit  0: getpid

After:
  # uftrace record --debug-domain plthook:3 t-abc.arm
      ...
  plthook: [idx:    2] enter 10480: getpid@plt (mod: 14b9e38)
  plthook: [idx:    2] exit  0: getpid

Since the address of enter actually points to the PLT address of the
function, it's better to add "@plt" postfix.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>