mcount: Fix mtdp assertion failure in plthook_exit
The mtd_key deletion in mcount_cleanup() sometimes makes other thread
get crashed because they may try to get mtdp based on this mtd_key after
being deleted.  The error message is as follows:

  plthook_exit: Assertion `!check_thread_data(mtdp)' failed.

The expected behavior is to call mcount_cleanup() after all the other
threads are destoryed, but it doesn't work as expected in some cases.

Since this key deletion is not that important, it'd be better not to do
it until we find a better solution.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>