tests: Add TestBase.is_32bit to check ELFCLASS
In some tests, special modification is needed for 32-bit environments
and it's being handled as follows:

  import platform
  if platform.architecture()[0].startswith('32bit'):
      # handle 32-bit specific manipulation

However, it doesn't work if user wants to run i386 tests under x86_64
environment using 32-bit compatible layer support.

So this patch introduces a correct way of checking whether the target
ELF image is 32-bit or not.  Instead of using platform.architecture(),
it can be used as follows:

  if TestBase.is_32bit(self):
      # handle 32-bit specific manipulation

It always works correctly by checking ELFCLASS of the target image.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>