tests: Fix 083 arg_float for 32-bit environments
The test t083_arg_float.py is incorrectly manipulated for 32-bit
environments.

  $ ./runtest.py -vdp -O0 083
      ...
  t083_arg_float: diff result of -pg -O0
  --- expect      2020-01-12 20:41:26.060842246 +0900
  +++ result      2020-01-12 20:41:26.060842246 +0900
  @@ -3,4 +3,4 @@
      float_sub(0.100000, 0.200000) = -0.100000;
  -   float_mul(300.000000, 400.000000) = 120000.000000;
  -   float_div(40000000000.000000, -0.020000) = -2000000000000.000000;
  +   float_mul(300.000000, 3.792969) = 120000.000000;
  +   float_div(40000000000.000000, 19991592893437257463236597020847243264.000000) = -2000000000000.000000;
    } /* main */

  083 arg_float           : NG

This patch fixes the arg number because the second argument has to
manualy be set to arg3 because uftrace currently calculates the offset
only with its index number.

It's user's responsibility to give the proper index number if DWARF info
is not given.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>