tests: Add TestBase.get_elf_machine() to runtests.py
It'd be better to check e_machine from the given ELF file rather than
checking the machine based on the running environment.

It's required because i386 ELF binary may be executed under x86_64
machine with compatibility layer.  In this case platform.machine()
always returns x86_64 and there's no way to check if the target ELF
binary is built in i386 mode.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>