script: Support native C/C++ shared object scripting
This patch adds a native script support that can use a shared object as
a script plugin.

The basic interface is same as python and luajit, but the benefit of
this native script is that it doesn't add any overhead from the
scripting engines.  It just bypasses the execution flow into the
dynamically loaded shared object.

The target shared object has to be compiled from C/C++ programs that
provides the following functions after including "include/script.h".

  void uftrace_begin(struct uftrace_script_info *sc_info);
  void uftrace_entry(struct uftrace_script_base_ctx *sc_ctx);
  void uftrace_exit(struct uftrace_script_base_ctx *sc_ctx);
  void uftrace_end(void);

The 'struct uftrace_script_base_ctx' is same as the struct used inside
uftrace so it's just passed without any manipulation.

The example C files are in sctipts directory and the shared objects can
be compiled as follows.

  $ make -C scripts/
  make: Entering directory '/home/honggyu/work/uftrace/scripts'
  gcc -fPIC -shared -g -o simple.so simple.c
  gcc -fPIC -shared -g -o info_c.so info.c
  gcc -fPIC -shared -g -o count.so count.c
  gcc -fPIC -shared -g -o dump.so dump.c
  make: Leaving directory '/home/honggyu/work/uftrace/scripts'

The basic usage is as follows.

  $ uftrace record -S scripts/simple_c.so --no-libcall t-abc
  program begins...
  entry : main()
  entry : a()
  entry : b()
  entry : c()
  exit  : c()
  exit  : b()
  exit  : a()
  exit  : main()
  program is finished

The output looks exactly same as python and luajit scripts.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>