script: Change script_info.cmds from struct strv to char*
It doesn't look like a good idea to expose 'struct strv' type in
'struct script_info' when it's exposed to external users.

Rather than using the internal data structure to represent commands,
it'd be much simpler to use a concatenated char* string.

So this patch changes the type of script_info.cmds from 'struct strv' to
'const char*' type as follows.

 struct script_info {
        char                    *name;
        char                    *version;
        bool                    record;
-       struct strv             cmds;
+       const char              *cmds;
 };

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>